Organize acceptance test code
